# backend/swagger/userRoutes.yml
openapi: 3.0.0
info:
  title: User API
  version: 1.0.0
paths:
  /signup:
    post:
      summary: Register a new user
      requestBody:
        description: User information for registration
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User registered successfully
  /login:
    post:
      summary: Log in as an existing user
      requestBody:
        description: User credentials for login
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User logged in successfully
  /{id}:
    delete:
      summary: Delete a user by ID
      parameters:
        - in: path
          name: id
          description: User ID to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
    put:
      summary: Update a user by ID
      parameters:
        - in: path
          name: id
          description: User ID to update
          required: true
          schema:
            type: string
      requestBody:
        description: Updated user information
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
      responses:
        '200':
          description: User updated successfully
    get:
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          description: User ID to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User retrieved successfully